+++ include/erl_sdf_mapping/gp_sdf_mapping_3d.hpp
@@ -99,8 +99,11 @@ namespace erl::sdf_mapping {
         std::mutex m_log_mutex_;                                                                // for logging
         double m_travel_distance_ = 0;                                                          // for logging
         std::optional<Eigen::Vector3d> m_last_position_ = std::nullopt;                         // for logging
-        std::ofstream m_train_log_file_;
-        std::ofstream m_test_log_file_;
+        std::ofstream m_train_log_file_;                                                        // for logging
+        std::ofstream m_test_log_file_;                                                         // for logging
+        std::size_t m_num_update_calls_ = 0;                                                    // for logging
+        std::size_t m_num_test_calls_ = 0;                                                      // for logging
+        std::size_t m_num_test_positions_ = 0;                                                  // for logging
 
         // for testing
         struct TestBuffer {
@@ -174,6 +177,21 @@ namespace erl::sdf_mapping {
         [[nodiscard]] const OctreeKeyGpMap&
         GetGpMap() const {
             return m_gp_map_;
+        }
+
+        [[nodiscard]] std::size_t
+        GetNumUpdateCalls() const {
+            return m_num_update_calls_;
+        }
+
+        [[nodiscard]] std::size_t
+        GetNumTestCalls() const {
+            return m_num_test_calls_;
+        }
+
+        [[nodiscard]] std::size_t
+        GetNumTestPositions() const {
+            return m_num_test_positions_;
         }
 
         [[nodiscard]] bool