+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2024-07-28
+
+- Add statics of update calls, test calls, and test positions
+
 # 2024-07-24
 
 - Update README.md